requirements.txt
# Document Image Denoising - Requirements

# Core dependencies
tensorflow>=2.15.0
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Optional (for better performance)
# tensorflow-directml-plugin  # For Windows GPU support